We do not support character buffer protocol on QStrings.
The only way of supporting it is storing an 8-bit encoded version of QString
into the QString instance, this will increase the memory usage of QString in 50%,
and we don't want it.

So, if you need to pass QStrings for a python functions which doesn't accept
unicode, just cast yourself the QString to a python string.