Fixes QtGui type system to use 'reference-count' tag setModel methods.
Various setModel methods where changed from the incorrect parenting
tag to the proper 'reference-count' tag. The return value of
QAbstractItemView.model() was defined to have 'default' ownership,
otherwise the Qt heuristics from the generator would set the returned
pointer as a child belonging to the QAbstractItemView instance.
The test case for this was expanded and changed to a more meaningful
name.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>