QEasingCurve::setCustomType typesystem workaround
Due to bug #201, typesystem signatures that take pointer to function
types are incorrectly expanded.

To workaround this bug, the typesystem signature for
QEasingCurve::setCustomType removal is:

<modify-function signature="setCustomType(double)" remove="all"/>

But for ARM, this same workaround does not work, and has to be:

<modify-function signature="setCustomType(float)" remove="all"/>

This is probably due to how apiextractor/shiboken is incorrectly
expanding the pointer to function signature. To remove setCustomType on
ARM (and thus allow the build to succeed), this commit adds both
signatures.

This is a temporary fix just while bug #201 is not fixed.

Reviewed-by: Lauro Moura <lauro.neto@openbossa.org>
Reviewed-by: Bruno Araujo <bruno.araujo@openbossa.org>