Replace macro checks with actual symbol checking
This increases a little more the check time, but is also more precise.
It is also required for cases where qfeatures.h incorrectly reports
support for a feature, but the actual class is not compiled (the case
for a couple of classes in Qt Simulator).

Additionally, the macro has been made more concise, and is reused for
both QtGui and QtNetwork.

Reviewed-by: Luciano Wolf <luciano.wolf@openbossa.org>
Reviewed-by: Hugo Parente Lima <hugo.lima@openbossa.org>