Fixed segmentation fault libpyside's signalUpdateSource function.
This function is called when an object is instantiated, and it will
go through the class attributes looking for signals and what else
is relevant. If the user has set a new attribute in the constructor
before the call to its parent QObject-like __init__ method, a
segmentation fault would ensue.

This commit fixes this condition and also adds an unit test.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>
Reviewed by Luciano Wolf <luciano.wolf@openbossa.org>