Added a PySide derived test library.
The test library is intended to provide specific test cases in a way
that problems could be isolated easier than using the ones provided by
the Qt library. Cases commons to all Python bindings must go into the
Shiboken test libraries, the one added here is intended to test Qt
specifics (e.g. signals).

A Python unit test was added for the case when a method and a signal
have the same name.