Fix BUG #1135 - "SIGSEGV when loading custom widget using QUiLoader..."
As PyCustomWidget::createWidget()'s caller is a thread which calls
the PyObject_CallObject() function it needs _exclusive_ access to the
Python interpreter so that a Shiboken::Gilstate is needed in order to
make it thread-safe and avoiding that other threads call the interpreter
at the same time.

See http://bugs.pyside.org/show_bug.cgi?id=1135.

Signed-off-by: Paulo Alcantara <pcacjr@zytor.com>