Do not always try to remove bytecode version of the test.
Trying to remove the bytecode file doesn't work if it wasn't created
in the first place.  This can happen for example under Gentoo, where
bytecode is disabled while building packages to avoid sandbox
violations when Python tries to write a bytecode file for an
already-installed module outside the package build directory.

Change-Id: I4cbd63527479365e6f201d3fc51c576f55830680
Reviewed-by: Hugo Parente Lima <hugo.lima@openbossa.org>