Use QObject property to invalidate wrapper before deletion.
Define a getWrapperForQObject function to create / find
wrappers for QObject derived objects.  It sets a property
so that an invalidate function can be called when the QObject
property is cleared in the QObject destructor after all destroyed
signal handlers have been called.  Requires a change to shiboken
to generate calls to the getWrapperForQObject function.

Change-Id: I4ed548af114bf7176e8e8255a3ea70549c35d74c
Reviewed-by: Hugo Parente Lima <hugo.lima@openbossa.org>