Fix PYSIDE-61
The original author of the test did not know how to create
one webview after the other after one webview finished loading,
so the old approach segfaulted, because it tried to access variables
that were already freed.

The new approach simply creates webviews with a timer after the
previous webview finished loading. It is also using UsesQApplication
instead of creatig the qApp manually.

Change-Id: I871e7a238398d96e110e89872634a9c5f3b5bc12
Reviewed-by: John Ehresman <jpe@wingware.com>